+++ CHANGELOG.md
@@ -3,3 +3,7 @@
 
 ## [1.0.1] - 2020-02-15
 * Final Version
+
+## [1.0.2] - 2020-08-15
+* Page Title and Description style customization
+* Possibility to change buttons text